pyproject.toml
[project]
name = "stock-agent"
version = "0.1.0"
description = "Python 한국주식 자동매매 시스템 (KIS Developers + ORB)"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "python-kis>=2.1.0",
    "python-telegram-bot>=21.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "loguru>=0.7",
    "pykrx>=1.0.45",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "black>=24.8",
    "pytest>=8.3",
    "pytest-mock>=3.14",
    "pre-commit>=3.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stock_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "scripts", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]

[tool.ruff.lint.per-file-ignores]
# 테스트 함수명에 한글 사용 허용 (가독성 우선).
"tests/**/*.py" = ["N802"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.pyright]
# nvim/VS Code 등의 pyright LSP 가 프로젝트 venv 와 src layout 을 인식하도록.
venvPath = "."
venv = ".venv"
include = ["src", "scripts", "tests"]
extraPaths = ["src"]
pythonVersion = "3.11"
reportMissingImports = "error"
